updated kanban
