--- test/chat/footer.vue
@@ -1,13 +1,0 @@
-<template>
-  <text-chat ref="textChat" />
-</template>
-
-<script>
-import TextChat from "./text-chat";
-export default {
-  name: "footer",
-  components: {
-    TextChat
-  }
-};
-</script>
